@yuantailing 2025-09-29
Yes, compiled cat/copy improves perf when layout is strided.

For `cat(a, b)`, if both a and b are contiguous, torch calls `CatArrayBatchedCopy_alignedK_contig` which is fast; if a or b is strided, torch calls `CatArrayBatchedCopy` which is slower than the compiled kernel.

For `a.copy_(b)`, direct_copy_kernel_cuda is always strided or torch will call cudaMemcpy.